Refactorizar codigo para cargar excel
