+++ harvester-app/src/framework/utils/css/moduloInicio.css
@@ -143,7 +143,7 @@
   align-items: center;
   justify-content: center;
 }
-.cargar-input{
+.cargar-excel{
     display: none;
 }
 .avanzar-analisis:disabled {
