+++ harvester-app/src/framework/vistas/ModuloInicio.html
@@ -19,7 +19,7 @@
                 </button>
             </div>
             <label class="cargar">
-              <input type="file" accept=".xlsx" class="cargar-input" />
+              <input type="file" accept=".xlsx" class="cargar-excel" />
               <div class="texto-cargar">Cargar Archivo</div>
             </label>
             <button disabled class="avanzar-analisis">
